refactor: 리뷰 결과 반영
리뷰 사항 중 간단 개선건들 반영
- util -> domain 이동
- OutputView로 네이밍 변경
- Controller에 있던 replaceAll을 InputView로 이동
- Stream 문법 적용
