+++ src/main/java/lotto/controller/LottoController.java
@@ -3,10 +3,10 @@ package lotto.controller;
 import lotto.domain.*;
 import lotto.domain.dto.StatisticDto;
 import lotto.domain.enums.Rank;
-import lotto.util.ManualLottoGenerator;
-import lotto.util.RandomLottoGenerator;
+import lotto.domain.ManualLottoGenerator;
+import lotto.domain.RandomLottoGenerator;
 import lotto.view.InputView;
-import lotto.view.ResultView;
+import lotto.view.OutputView;
 
 import java.util.Arrays;
 import java.util.List;
@@ -15,13 +15,11 @@ import java.util.stream.Collectors;
 
 public class LottoController {
     private static final Pattern STANDARD_PATTERN = Pattern.compile(",");
-    private static final String WHITE_SPACE_REGEX_STRING = "\\s";
-    private static final String EMPTY = "";
 
     public void run() {
         Money money = new Money(InputView.getMoney());
         List<LottoTicket> tickets = LottoMarket.sell(money, new RandomLottoGenerator());
-        ResultView.printTickets(tickets);
+        OutputView.printTickets(tickets);
 
         showLottoResult(money, tickets);
     }
@@ -32,15 +30,12 @@ public class LottoController {
         );
         StatisticDto dto = committee.statistics(tickets);
 
-        ResultView.printStatistics(dto);
-        ResultView.printReturnRate(committee.returnRate(Rank.calculatePrice(dto), money));
+        OutputView.printStatistics(dto);
+        OutputView.printReturnRate(committee.returnRate(Rank.calculatePrice(dto), money));
     }
 
     private List<Integer> createWinningNumbers() {
-        String input = InputView.getWiningNumber()
-                .replaceAll(WHITE_SPACE_REGEX_STRING, EMPTY);
-
-        return Arrays.asList(STANDARD_PATTERN.split(input)).stream()
+        return Arrays.stream(STANDARD_PATTERN.split(InputView.getWiningNumber()))
                 .map(Integer::parseInt)
                 .collect(Collectors.toList());
     }
