+++ src/main/java/lotto/domain/LottoCommittee.java
@@ -13,9 +13,7 @@ public class LottoCommittee {
 
     public StatisticDto statistics(List<LottoTicket> tickets) {
         StatisticDto dto = StatisticDto.create();
-        for (LottoTicket ticket : tickets) {
-            dto.add(countOfMatch(ticket));
-        }
+        tickets.forEach(ticket -> dto.add(countOfMatch(ticket)));
 
         return dto;
     }
