+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,6 +1,4 @@
-package lotto.util;
-
-import lotto.domain.LottoTicket;
+package lotto.domain;
 
 public interface LottoGenerator {
     LottoTicket create();
