+++ src/main/java/lotto/view/InputView.java
@@ -4,9 +4,13 @@ import java.util.Scanner;
 
 public class InputView {
     private static final String INVALID_NUMBER_MESSAGE = "입력값이 숫자가 아닙니다.";
+    private static final String WHITE_SPACE_REGEX_STRING = "\\s";
+    private static final String EMPTY = "";
+
+    private InputView() {}
 
     public static int getMoney() {
-        ResultView.printGetMoney();
+        OutputView.printGetMoney();
 
         try {
             return Integer.parseInt(read());
@@ -16,8 +20,8 @@ public class InputView {
     }
 
     public static String getWiningNumber() {
-        ResultView.printWinningNumber();
-        return read();
+        OutputView.printWinningNumber();
+        return read().replaceAll(WHITE_SPACE_REGEX_STRING, EMPTY);
     }
 
     private static String read() {
