+++ src/main/java/lotto/view/OutputView.java
@@ -6,7 +6,9 @@ import lotto.domain.dto.StatisticDto;
 
 import java.util.List;
 
-public class ResultView {
+public class OutputView {
+    private OutputView() {}
+
     public static void printGetMoney() {
         System.out.println("구입금액을 입력해 주세요.");
     }
