+++ src/test/java/lotto/domain/LottoCommitteeTest.java
@@ -2,7 +2,6 @@ package lotto.domain;
 
 import lotto.domain.dto.StatisticDto;
 import lotto.fixture.LottoTicketFixture;
-import lotto.util.ManualLottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
