+++ src/test/java/lotto/domain/LottoMarketTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.util.RandomLottoGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
